requirements.txt
Flask
google-cloud-speech
googletrans==4.0.0-rc1
pydub